feature: implement MembersPage component to display group members and their details
